Core tests v10
FAILED (failures=5, errors=15, skipped=3)
