+++ job_connect/core/tests/_test_helpers.old
@@ -1,0 +1,7 @@
+import unittest
+from django.test import TestCase, Client
+from django.contrib.auth import get_user_model
+
+User = get_user_model()
+
+
