+++ job_connect/core/urls.py
@@ -1,6 +1,8 @@
 from django.urls import path
 from . import views
 
+app_name = 'core'
+
 urlpatterns = [
     path('signup/applicant/', views.applicant_signup, name='applicant_signup'),
     path('signup/recruiter/', views.recruiter_signup, name='recruiter_signup'),
