+++ job_connect/core/views.py
@@ -23,7 +23,7 @@ def applicant_signup(request):
                 user.groups.add(applicant_group)
 
             login(request, user)
-            return redirect('applicant/applicant_profile_create')  # Redirect to create profile
+            return redirect('applicant:applicant_profile_create')  # Redirect to create profile
     else:
         form = ApplicantSignUpForm()
     return render(request, 'core/applicant_signup.html', {'form': form})
@@ -33,18 +33,18 @@ def recruiter_signup(request):
         form = RecruiterSignUpForm(request.POST)
         if form.is_valid():
             user = form.save()
-            # Assign the user to the 'Applicant' group
+            # Assign the user to the 'Recruiter' group
             try:
                 recruiter_group = Group.objects.get(name='Recruiter')
                 user.groups.add(recruiter_group)
             except Group.DoesNotExist:
-                print("Error: Applicant group does not exist!")
+                print("Error: Recruiter group does not exist!")
                 # Create the group if it doesn't exist:
                 recruiter_group = Group.objects.create(name='Recruiter')
                 user.groups.add(recruiter_group)
 
             login(request, user)
-            return redirect('recruiter/recruiter_profile_create') # Redirect to create profile
+            return redirect('recruiter:recruiter_profile_create') # Redirect to create profile
     else:
         form = RecruiterSignUpForm()
     return render(request, 'core/recruiter_signup.html', {'form': form})
@@ -56,9 +56,9 @@ def login_view(request):
             user = form.get_user()
             login(request, user)
             if user.is_applicant():
-                return redirect('applicant_dashboard') # Define applicant dashboard URL
+                return redirect('applicant:applicant_dashboard') # Define applicant dashboard URL
             elif user.is_recruiter():
-                return redirect('recruiter_dashboard') # Define recruiter dashboard URL
+                return redirect('recruiter:recruiter_dashboard') # Define recruiter dashboard URL
             else:
                 return redirect('home') # Default redirect
     else:
