+++ job_connect/job/urls.py
@@ -1,6 +1,8 @@
 from django.urls import path
 from . import views
 
+app_name = 'job'
+
 urlpatterns = [
     path('', views.JobListView.as_view(), name='job_list'),
     path('search/', views.JobSearchView.as_view(), name='job_search'),
